+++ Documentation/manual/CHANGELOG.md
@@ -31,6 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - <code>DrawColliderGizmo</code> no longer throws a null ref when not given a collider.
 - <code>MenuPageInspector</code> no longer fuses the path name and file name when creating a <code>MenuState</code> with the inspector button.
 - Possible null refs in <code>ApplicationStateMachine</code> when not supplied with play or pause signals.
+- <code>RouterButton</code> now correctly removes level load listeners instead of adding more... <sub>god damn typos</sub>.
+- <code>RouterButton</code> now caches its button in <code>OnValidate</code> even if it is not null.
+  - This change fixes a bad reference that occurs when copy/pasting the component.
 
 ## [v0.0.0-alpha.10.4]